Finalize all the things
