+++ src/fXmlRpc/Exception/HttpException.php
@@ -25,7 +25,7 @@ namespace fXmlRpc\Exception;
 
 use Exception;
 
-class HttpException extends TransportException
+final class HttpException extends TransportException
 {
     public static function httpError($message, $statusCode, Exception $previous = null)
     {
