+++ src/fXmlRpc/Exception/RuntimeException.php
@@ -25,29 +25,6 @@ namespace fXmlRpc\Exception;
 
 use RuntimeException as BaseRuntimeException;
 
-class RuntimeException extends BaseRuntimeException implements ExceptionInterface
+abstract class RuntimeException extends BaseRuntimeException implements ExceptionInterface
 {
-    public static function unexpectedTag($tagName, $elements, array $definedVariables, $depth, $xml)
-    {
-        $expectedElements = [];
-        foreach ($definedVariables as $variableName => $variable) {
-            if (substr($variableName, 0, 4) !== 'flag') {
-                continue;
-            }
-
-            if (($elements & $variable) === $variable) {
-                $expectedElements[] = substr($variableName, 4);
-            }
-        }
-
-        return new static(
-            sprintf(
-                'Invalid XML. Expected one of "%s", got "%s" on depth %d (context: "%s")',
-                implode('", "', $expectedElements),
-                $tagName,
-                $depth,
-                $xml
-            )
-        );
-    }
 }
